+++ src/tests/end2end/InputStateTests.cpp
@@ -138,7 +138,7 @@ class InputStateTest : public DawnTest {
         struct AttributeSpec {
             uint32_t location;
             uint32_t slot;
-            uint32_t offset;
+            uint64_t offset;
             VertexFormat format;
         };
         dawn::InputState MakeInputState(std::vector<InputSpec> inputs, std::vector<AttributeSpec> attributes) {
@@ -183,7 +183,7 @@ class InputStateTest : public DawnTest {
             dawn::RenderPassEncoder pass = encoder.BeginRenderPass(&renderPass.renderPassInfo);
             pass.SetPipeline(pipeline);
 
-            uint32_t zeroOffset = 0;
+            uint64_t zeroOffset = 0;
             for (const auto& buffer : vertexBuffers) {
                 pass.SetVertexBuffers(buffer.location, 1, buffer.buffer, &zeroOffset);
             }
@@ -456,7 +456,7 @@ TEST_P(InputStateTest, UnusedVertexSlot) {
 
     dawn::RenderPassEncoder pass = encoder.BeginRenderPass(&renderPass.renderPassInfo);
 
-    uint32_t zeroOffset = 0;
+    uint64_t zeroOffset = 0;
     pass.SetVertexBuffers(0, 1, &buffer, &zeroOffset);
     pass.SetVertexBuffers(1, 1, &buffer, &zeroOffset);
 
@@ -501,7 +501,7 @@ TEST_P(InputStateTest, MultiplePipelinesMixedInputState) {
 
     dawn::RenderPassEncoder pass = encoder.BeginRenderPass(&renderPass.renderPassInfo);
 
-    uint32_t zeroOffset = 0;
+    uint64_t zeroOffset = 0;
     pass.SetVertexBuffers(0, 1, &buffer, &zeroOffset);
     pass.SetVertexBuffers(1, 1, &buffer, &zeroOffset);
